pyproject.toml
[tool.poetry]
name = "gas_split"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.0"
uvicorn = "^0.14.0"
python-multipart = "^0.0.5"
Jinja2 = "^3.0.1"
aiofiles = "^0.7.0"
shortuuid = "^1.0.1"
pymongo = {extras = ["srv"], version = "^3.12.0"}

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
python-dotenv = "^0.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
